LayoutTests/fast/frames/seamless/seamless-quirks-expected.txt

Test that a quirks-mode document inside a seamless iframe sizes correctly.
FAIL iframe.seamless should be true (of type boolean). Was undefined (of type undefined).
PASS iframe.contentDocument.compatMode is "BackCompat"
PASS window.getComputedStyle(iframe).width is "200px"
PASS window.getComputedStyle(iframe).height is "100px"
